+++ pages/markdown/misc/changelog.md
@@ -2,6 +2,7 @@
 * **Fixed Flashcard Divider Line:** Fixed the line that divides the left and right of flashcards in a list to be the full height
 * **Fixed Flashcard Number Display:** Fixed flashcards' displayed numbers to now properly increment with each flashcard
 * **Fixed Flashcard List Order:** Fixed flashcard list order to be based on the flashcards' subsection and mainsection, rather than simply the order number
+* **Improved Registering:** Improved autocomplete functionality and added a strength bar to the password field
 
 ## Shared Deck Hotfix - Dec 08 - 1.0.11
 Quick hotfix for shared decks.